+++ resources/assets/js/alpine-directives/dialog.js
@@ -19,6 +19,24 @@ const Dialog = (config) => ({
                 });
             }
         });
+
+        window.addEventListener('close-dialog', (e) => {
+            const firstChild = this.$el.firstElementChild;
+
+            if (!firstChild) {
+                throw new Error('Dialog component should be wired or must have a child element with an id attribute.');
+            }
+
+            if (firstChild.id === e.detail.id) {
+                this.close();
+
+                this.$dispatch('dialog-closed', {
+                    id: e.detail.id,
+                    el: this.$el,
+                    trigger: e.target,
+                });
+            }
+        });
     },
     open() {
         this.isOpen = true;