+++ src/Table/UI/views/columns/header.blade.php
@@ -1,5 +1,5 @@
-<th x-show="selection.length == 0" scope="col" class="text-left">
+<th wire:key="{{ $getKey() }}" wire:ignore x-show="selection.length == 0" scope="col" class="text-left">
     <span class="text-xs/5 font-medium uppercase tracking-wider text-grey-700">
-        {{ $getLabel }}
+        {{ $getLabel() }}
     </span>
 </th>